# Backend/jiufang-project/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: 九方智能查询系统 API
  description: |
    九方智能查询系统是一个基于AI Agent的自然语言查询系统，支持用户通过自然语言查询ERP数据，
    并提供定时报告、预警规则、数据导出等功能。
    
    ## 认证方式
    - 使用JWT Bearer Token认证
    - 登录接口获取JWT令牌
    - 令牌有效期24小时
    
    ## 权限控制
    - admin: 系统管理员，拥有所有权限
    - manager: 管理层，拥有查询和导出权限
    - executive: 执行层，拥有查询权限
    
    ## 响应格式
    所有接口统一使用以下响应格式：
    ```json
    {
      "code": 200,
      "message": "success",
      "data": {}
    }
    ```
  version: 1.0.0
  contact:
    name: 九方技术团队
    email: tech@jiufang.com
servers:
  - url: https://api.jiufang.com/v1
    description: 生产环境
  - url: https://api-dev.jiufang.com/v1
    description: 开发环境
  - url: http://localhost:8080/api/v1
    description: 本地开发环境
...
paths:
...
  /groups:
    post:
      tags:
        - Group
      summary: 创建用户组
      description: 创建新用户组（仅系统管理员可操作）
      operationId: createGroup
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                description:
                  type: string
                  maxLength: 500
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    get:
      tags:
        - Group
      summary: 查询用户组列表
      description: 分页查询用户组列表（仅系统管理员可操作）
      operationId: listGroups
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: name
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /groups/{group_id}:
    get:
      tags:
        - Group
      summary: 查询用户组详情
      description: 查询单个用户组详情（仅系统管理员可操作）
      operationId: getGroup
      security:
        - bearerAuth: []
      parameters:
        - name: group_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^\d{19}$'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Group
      summary: 更新用户组
      description: 更新用户组信息（仅系统管理员可操作）
      operationId: updateGroup
      security:
        - bearerAuth: []
      parameters:
        - name: group_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^\d{19}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                description:
                  type: string
                  maxLength: 500
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Group
      summary: 删除用户组
      description: 删除用户组（仅系统管理员可操作）
      operationId: deleteGroup
      security:
        - bearerAuth: []
      parameters:
        - name: group_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^\d{19}$'
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /groups/{group_id}/permissions:
    post:
      tags:
        - Group
      summary: 配置用户组权限
      description: 配置用户组权限（仅系统管理员可操作）
      operationId: configurePermissions
      security:
        - bearerAuth: []
      parameters:
        - name: group_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^\d{19}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - permissions
              properties:
                permissions:
                  type: array
                  items:
                    type: object
                    required:
                      - table_name
                      - allowed_fields
                    properties:
                      table_name:
                        type: string
                        minLength: 1
                        maxLength: 100
                      allowed_fields:
                        type: string
                        description: JSON数组格式
                      filter_condition:
                        type: string
      responses:
        '200':
          description: 配置成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    get:
      tags:
        - Group
      summary: 查询用户组权限
      description: 查询用户组权限配置（仅系统管理员可操作）
      operationId: getPermissions
      security:
        - bearerAuth: []
      parameters:
        - name: group_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^\d{19}$'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /groups/{group_id}/members:
    get:
      tags:
        - Group
      summary: 查询用户组成员
      description: 查询用户组成员列表（仅系统管理员可操作）
      operationId: getGroupMembers
      security:
        - bearerAuth: []
      parameters:
        - name: group_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^\d{19}$'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupMemberListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags:
        - Group
      summary: 添加用户组成员
      description: 添加用户组成员（仅系统管理员可操作）
      operationId: addGroupMembers
      security:
        - bearerAuth: []
      parameters:
        - name: group_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^\d{19}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_ids
              properties:
                user_ids:
                  type: array
                  items:
                    type: string
                    pattern: '^\d{19}$'
      responses:
        '200':
          description: 添加成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /groups/{group_id}/members/{user_id}:
    delete:
      tags:
        - Group
      summary: 删除用户组成员
      description: 删除用户组成员（仅系统管理员可操作）
      operationId: removeGroupMember
      security:
        - bearerAuth: []
      parameters:
        - name: group_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^\d{19}$'
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^\d{19}$'
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== 历史记录模块 ====================
...
components:
  schemas:
    SuccessResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: 错误码
          example: 40001
        message:
          type: string
          description: 错误信息
          example: "参数错误"
...
    GroupResponse:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          required:
            - id
            - name
            - created_at
          properties:
            id:
              type: string
              pattern: '^\d{19}$'
            name:
              type: string
            description:
              type: string
            created_at:
              type: string
              format: date-time
    GroupListResponse:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          required:
            - total
            - page
            - page_size
            - groups
          properties:
            total:
              type: integer
            page:
              type: integer
            page_size:
              type: integer
            groups:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                    pattern: '^\d{19}$'
                  name:
                    type: string
                  description:
                    type: string
                  created_at:
                    type: string
                    format: date-time
    PermissionListResponse:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          required:
            - permissions
          properties:
            permissions:
              type: array
              items:
                type: object
                properties:
                  table_name:
                    type: string
                  allowed_fields:
                    type: string
                  filter_condition:
                    type: string
    GroupMemberListResponse:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          required:
            - members
          properties:
            members:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                    pattern: '^\d{19}$'
                  username:
                    type: string
                  email:
                    type: string
                    format: email
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 40001
            message: "参数错误"
    Unauthorized:
      description: 未授权
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 40101
            message: "未授权访问"
    Forbidden:
      description: 权限不足
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 40301
            message: "权限不足"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 40401
            message: "资源不存在"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT认证